+++ packages/moonstone/CHANGELOG.md
@@ -53,6 +53,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Notification` to break line in characters in ja and zh locale
 - `moonstone/Notification` to align texts left in LTR locale and right in RTL locale
 - `moonstone/VideoPlayer` to simulate rewind functionality on non-webOS platforms only
+- `moonstone/Panels` to pass the `onTransition` prop to its underlying `<Viewport>`
 
 ### Fixed